feat:
The purpose of rededign logicblock-legacy is to make it less coupling.
But I realized that, exporting a Board has to do with dealing a dependency cycle.
A possible solution is adding aspects in blocks, but that's too coupling, and that's what I have done in logicblock-legacy.
I haven't found any non-invasive solution so far.